Add types for custom adapter and remove ts-ignore
