+++ types/types.d.ts
@@ -4,6 +4,8 @@ import type {
 	User as OGUser,
 } from '@auth/sveltekit/node_modules/@auth/core/types';
 import type { UserSettings } from '@prisma/client';
+import { SvelteKitAuthConfig as OGSvelteKitAuthConfig } from '@auth/sveltekit';
+import type { CustomAdapter } from '$lib/prisma/client';
 
 // TODO: update when they fix this:
 // https://github.com/nextauthjs/next-auth/issues/6640#issuecomment-1426801813
@@ -17,5 +19,11 @@ declare module '@auth/sveltekit/node_modules/@auth/core/types' {
 
 	interface User extends OGUser {
 		settings: UserSettings;
+	}
+}
+
+declare module '@auth/sveltekit' {
+	interface SvelteKitAuthConfig extends OGSvelteKitAuthConfig {
+		adapter: CustomAdapter;
 	}
 }
